moved event handling code to own file
